+++ src/Parser/Structure/Record/Filter/ColorMatrixFilter.php
@@ -21,7 +21,6 @@ declare(strict_types=1);
 namespace Arakne\Swf\Parser\Structure\Record\Filter;
 
 use Arakne\Swf\Parser\SwfReader;
-
 use Override;
 
 use function assert;