+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+0.2.1
+-----
+
+- (Parser) Fix invalid end of stream detection during uncompress SWF file
+- (Parser) Use `EXTRA_DATA` error instead of `INVALID_DATA` when uncompressed data is larger than expected
+- Enable PHP-CS-Fixer for the project
+
 0.2.0
 -----